Separate token middleware for case routes
