+++ backend/src/app.module.ts
@@ -37,6 +37,7 @@ import { EntityCodeModule } from './v1/entity_code/entity_code.module';
 import { StagingMetaDataMappingModule } from './v1/staging_meta_data_mapping/staging_meta_data_mapping.module';
 import { StagingStatusCodeModule } from './v1/staging_status_code/staging_status_code.module';
 import { StagingActivityCodeModule } from './v1/staging_activity_code/staging_activity_code.module';
+import { RequestTokenMiddleware } from "./middleware/req.token";
 
 console.log("Var check - POSTGRESQL_HOST", process.env.POSTGRESQL_HOST);
 console.log("Var check - POSTGRESQL_DATABASE", process.env.POSTGRESQL_DATABASE);
@@ -106,5 +107,6 @@ export class AppModule {
   // let's add a middleware on all routes
   configure(consumer: MiddlewareConsumer) {
     consumer.apply(HTTPLoggerMiddleware).forRoutes("*");
+    consumer.apply(RequestTokenMiddleware).forRoutes("v1/code-table", "v1/case/*");
   }
 }
