+++ backend/src/middleware/req.res.logger.ts
@@ -1,5 +1,5 @@
 import {Request, Response, NextFunction} from 'express';
-import {Injectable, NestMiddleware, Logger, HttpException, HttpStatus} from '@nestjs/common';
+import {Injectable, NestMiddleware, Logger} from '@nestjs/common';
 
 @Injectable()
 export class HTTPLoggerMiddleware implements NestMiddleware {
@@ -8,22 +8,11 @@ export class HTTPLoggerMiddleware implements NestMiddleware {
   use(request: Request, response: Response, next: NextFunction): void {
     const {method, originalUrl} = request;
 
-    const authHeader = request.get('authorization');
-    if (!authHeader) {
-      throw new HttpException('No auth token', HttpStatus.UNAUTHORIZED);
-    }
-    const bearerToken: string[] = authHeader.split(' ');
-    const token: string = bearerToken[1];
-
-    //@ts-ignore
-    request.token  = token;
-
     response.on('finish', () => {
       const {statusCode} = response;
       const hostedHttpLogFormat = `${method} ${originalUrl} ${statusCode} - ${request.get('user-agent')}`;
       this.logger.log(hostedHttpLogFormat);
     });
-
     next();
   }
 }
